Implemented `GetValidatedMappingType()` for collection mapping
The `PopulateTargetCollectionAsync()` method takes topics from a collection on the source topic and maps them to models in a collection property on the target model. If the mapped topic type isn't compatible with the collection property type, then it is excluded, as a way of filtering out incompatible types.

By incorporating the new `GetValidatedMappingType()` method (88c7fbaa), we pre-validate the type compatibility, instead of requiring that the model first be mapped—a potentially expensive operation.

This satisfies relationships, children, and other collections, as called for by #83.