Implemented `GetValidatedMappingType()` for topic references
The `GetTopicReferenceAsync()` method takes a topic reference from the source topic and maps it to a property on the target model. If the mapped topic type isn't compatible with the target property type, then it is excluded, as a way of filtering out incompatible types.

By incorporating the new `GetValidatedMappingType()` method (88c7fbaa), we pre-validate the type compatibility, instead of requiring that the model first be mapped—a potentially expensive operation.

This satisfies the topic reference requirement called for in #83.