Ensure that `RemoveItem()` enforces business logic
Previously, when deleting items directly via `Remove()`, the `RemoveItem()` wouldn't enforce business logic. This is problematic because sometimes properties either a) disallow nulls, or b) must update their internal states when values are set to null. This would be handled when calling `SetValue()` or going directly through the property, but bypassed when calling `Remove()`.

This update patches that hole, and thus resolves the `Remove()` portion of #79.