Ensure that `ClearItems()` enforces business logic
Previously, when deleting items directly via `Clear()`, the `ClearItems()` wouldn't enforce business logic. This is problematic because sometimes properties either a) disallow nulls, or b) must update their internal states when values are set to null. This would be handled when calling `SetValue()` or going directly through the property, but bypassed when calling `Clear()`.

Unfortunately, for `Clear()`, this means checking the business logic for each individual record. As such, instead of just bulk deleting these, we're instead calling `Remove()` for each individual item. This is less performant, but ensures that the business logic is always accounted for. Fortunately, calling `Clear()` on this collection is uncommon; it's far more common to work with individual items via `SetValue()` or, less commonly, `Remove()`.

This update patches that hole, and thus resolves the `Clear()` portion of #79.