Updated tests which expected `SetValue(key, null)` to maintain record
Previously, calling `SetValue(key, null)` would set the `TrackedRecord<T>.Value` to null, but wouldn't delete the record. This was confusing, if not buggy, and the behavior was fixed (3d1ad6bc, 0bccf1f3). In doing so, however, unit tests which validated the previous assumption broke, and needed to be updated.

This isn't expected to harm any actual implementations, as they should already be written to account for missing records, since this is the default when loading from the database. The null value scenario would only occur when updating previously saved values.