Introduced test to ensure `Clear()`, `Remove()` enforce business logic
Introduced a test to ensure that calling `Clear()` enforces business logic by adding a property that will throw an exception if its set to null. This validates the behavior introduced in abfe213f. And since `Clear()` now calls `Remove()`, it also effectively validates the behavior introduced in 0bccf1f3. This validates the new functionality introduced as part of #79.