Introduced special `ModelType.Reflexive` enum value
The `Reflexive` value instructs callers of an `AttributeDescriptor` to gets its `ModelType` from the `AttributeDescriptor.ModelType` of the `AttributeDescriptor` derivative that the `AttributeDescriptor` is placed on.

This is a special scenario which allows e.g. the OnTopic Editor to display an attribute on an `AttributeDescriptor` which uses the current topic's values, and stores the result in the same location as the current topic.

This is useful for e.g. `AttributeDescriptor.DefaultValue`, which should be set to use the same attribute type and storage location as the `AttributeDescriptor` it is set for (OnTopicCMS/OnTopic-Editor-AspNetCore#44).