If no updates during `SetValue()`, return
If a call to `SetValue()` doesn't have `markDirty` or `version` set, and an existing `TrackedRecord<>` exists with the same `Value`, then don't create and store a new record; instead, just return, as there's nothing to update here. This is a minor performance improvement, since there's no reason to create a new instance of a `TrackedRecord<>` here. Notably, this impacts `Refresh()`, `Rollback()`, and the OnTopic Editor.