Bug fix: Equality comparison of values
For reasons that aren't entirely clear, the `==` check isn't using the value comparison when both `value` and `TrackedRecord<String>.Value` are `string`. This is unexpected. As a result, the values are _always_ being seen as having been changed, even when they haven't been. I'd like to investigate this further, as it could suggest behavior that will introduce bugs elsewhere in the library. In the meanwhile, I'm explicitly calling `Equals()` which remedies the problem, and resolves #86.