Established new `MemberAccessor` class for efficient reflection
When getting or setting e.g. property values via reflection, we're currently doing this using e.g. `PropertInfo.GetValue()` or `PropertInfo.SetValue()`. This is slow, as it first needs to dynamically generate a delegate. Since we have some knowledge of what types of members we'll be calling and what their signatures look like ahead of time, we can generate our own delegate, and then cache objects with these delegates, thus improving performance significantly. The `MemberAccessor` does this.

This is the first step toward implementation. Currently, the `MemberDispatcher` still calls reflection directly. This class will later be refactored into a set of extension methods which will operate off of `MemberAccessor`.