Established basic unit tests for the new `MemberAccessor`
These tests ensures that basic access to the `MemberAccessor` functions, including the validation and parsing of the required `MemberInfo`, as well as basic `GetValue()` and `SetValue()` methods.

This includes a new `MemberAccessorViewModel` for evaluating `MemberInfo` types specific to this class.

In practice, most access will be performed via higher level classes and, as such, we'll expect more comprehensive tests related to type compatibility and conversion to be performed as part of those classes in the future.