Established new `TypeAccessor` class for collecting `MemberAccessor`s
A `MemberAccessor` represents an individual member of a type. Typically, when using reflection to get or set values on an object, we'll be working against multiple members on the same type. To simplify this, the `TypeAccessor` will automatically establish a `MemberAccessor` for each compliant member (i.e., properties, methods with a return type and no parameters, and void methods with one parameter). It also includes passthrough methods to methods on `MemberAccessor` so that members can be validated, accessed, and updated all from `TypeAccessor`.