Established new `TypeAccessorFixture` for XUnit
In preparation for introducing unit tests for the `TypeAccessor`, the `TypeAccessorFixture` will allow for a persistent instance of a `TypeAccessor` for a given `Type` to be shared across multiple tests. Since individual `MemberAccessor`s are potentially expensive to create the first time, caching the entire collection of members for a given test type should reduce the time required for those tests.