Established basic unit tests for the new `TypeAccessor`
These tests ensures that basic access to the `TypeAccessor` (f246997f), including the validation and parsing of the required `Type`, the first-class `GetMembers()` and `GetMember()` methods, as well as the convenience passthrough methods to `MemberAccessor`, such as `HasGetter()`, `HasSetter()`, `GetValue()` and `SetValue()`.

Technically, the convenience methods are a bit redundant with the `MemberAccessor` tests. However, because the `TypeAccessor` tests operate against an entire type, this makes it easier for us to stand up `[Theory]` tests that evaluate multiple members and values at the same time. As a result, the tests for `GetValue()` and `SetValue()` are actually more comprehensive here than they are on the `MemberAccessor` tests.

This takes advantages of the new `TypeAccessFixture` (12716ed7) to share a single instance of a `TypeAccessor` across all tests.