Established new `TypeAccessorCache` class for caching `TypeAccessor`s
Each `MemberType` (eb7fc4a5) is expensive to create due to the overhead of establishing a delegate. If that is cached, however, it is dramatically faster than relying on reflection to dynamically create a delegate with each request. For that purpose, the `TypeAccessor` was setup to collect `MemberType` instances (f246997f), and now the `TypeAccessorCache` provides `static` access to `TypeAccessor` instances across the entire application, thus allowing `TypeAccessor` instances to be loaded once and accessed from anywhere.

This will replace the previous usage of private caches for e.g., `TopicPropertyDispatcher`, `TopicMappingService`, and `ReverseTopicMappingService`.