Updated exception type
In the unexpected case that a value being set to a property is not compatible with the target type, an `ArgumentException` is thrown by the reflection library. When using a delegate, the `InvalidCastException` is instead thrown. As this is a low-level, unexpected error that reflect a design flaw in e.g. a view model, it's not a concern that this changed (i.e., it's not worth rethrowing it as a `ArgumentException`). As such, I'm just updating the test to comply with the expectation.