Ensure member names are unique
This maintains consistency with the previous `MemberInfoCollection`, which also ensured uniqueness when adding members. Typically this isn't an issue, except in cases where e.g. a `new` member is added, hiding an inherited member. In that case, we only want the first result returned—which should be the `new` version. This is an outside use case that's not a primary concern, but we obviously want to avoid duplicate key exceptions being thrown in this scenario.