Migrated `MemberDispatcher` to `TypeAccessor` extensions
In a previous update, I retrofitted `MemberDispatcher` to operate off of `TypeAccessor` (and the underlying `MemberDispatcher`) instead of relying on `(Type)MemberInfoCollection` and going directly to the .NET Reflection classes when necessary.

With this done, it makes more sense for these methods to be treated as extensions of the `TypeMember` class—since they all rely upon it—instead of instantiating a new class that goes through the `TypeAccessorMemberCache`. This also provides a more intuitive naming scheme, since the legacy `MemberDispatcher` was a bit ambiguous with the new `MemberAccessor`.

The continues to maintain a separation between properties and methods. As this is no longer needed in the underlying library, we may want to revisit this at some point, and especially since these aren't currently validated as properties or methods. We can revisit this later.