Updated unit tests to account for new `TypeAccessor` extensions
With the `MemberDispatcher` refactored as extensions off of `TypeAccessor`, the legacy `TypeMemberInfoCollectionTest`—which, confusingly, primarily evaluated the `MemberDispatcher` class—has been renamed to `TypeAccessorExtensionsTest`, and updated to work off of that new class.

As part of this, I removed the `GetMember()` tests as this method is no longer used, and doesn't necessitate tests.

Most of these migrations were straight-forward updates to the new syntax, and have no functional impact. Much of the change in code is actually just rewrapping the XML docs.