Introduced required `allowConversion` parameter for `Set*Value()` methods
Previously, the `SetMethodValue()` and `SetPropertyValue()` _implicitly_ converted string values to compatible types (such as `int`, `bool`, and `DateTime`). This can potentially lead to unexpected scenarios if unexpected. While this is the expected—and long established—behavior, I decided to make it opt-in since it's something that should be a conscious decision by a developer to support.

This also allows the `SetPropertyValue()` and `SetMethodValue()` methods to optionally operate as faster, less-presumptuous, lower level operations, if needed. That will be supported in a subsequent update which moves most of this logic down to `MemberAccessor`'s `SetValue()` method.