Reverted expected exception back to `ArgumentException`
With the migration of the business logic back down to the lower-level `MemberAccessor`'s `SetValue()` method (a43434dc), we are now able to throw the previously expected `ArgumentException` (via `Contract.Assume()`) for the rare case where the object doesn't match the `MemberAccessor.MemberInfo.DeclaringType`.

This undoes one of the updates made as part of 56792e95.