Simplified `SetPropertyAsync()` logic
Since the `SetValue()` method now handles a lot of the business logic, we can simplify the logic for `SetPropertyAsync()`'s calls by no longer needing the validate compatibility first via the `TypeAccessor.HasSettableProperty()`.

All of this will just be repeated in `SetValue()` anyway—and, by design, will fail silently if it's not compatible—so we can just to a basic check against `CanWrite` to confirm it's an appropriate path.

(Even this last check isn't strictly necessary; this could be an `else` statement without any further conditions. But as these are quick checks, it helps reinforce what values won't get written at this higher level, and is thus more intuitive to read.)