Ensured member types are validated by `Has*ettable*()` methods
`HasGettableProperty()`, `HasGettableMethod()`, `HasSettableProperty()`, and `HasSettableMethod()` specify `Property` or `Method` as part of their signature. This isn't strictly necessary, as the underlying `GetValue()` and `SetValue()` on `MemberAccessor` will work against _either_ a property _or_ a method. As such, it remains a bit unclear—outside of semantic precision—whether we still need to maintain these overloads. Until then, however, they should validate what they claim to validate. Previously, if a gettable method name was passed to `HasGettableProperty()`, for instance, it would return true, since there was an `IsGettable` `MemberAccessor` with that name.  With this update, it explicitly verifies that it is, in fact, a property or a method.