Extended member type validation into `Set*Value()` methods
In a previous update, I updated the `Has*ettable*()` methods (e.g., `HasSettableProperty()` to validate that the member type is, in fact, a property or a method (44ead0eb). I've extended this validation to the `*et*Value()` methods (e.g., `SetPropertyValue()`) as well. This isn't technically required, but so long as we're maintaining the superficial difference between properties and methods, this distinction should be enforced.

Note that the setter will throw an exception if it's not found, whereas the getter will fail silently.