Removed unused `targetType`, `attributeFlag` overloads
The ability to validate the `targetType` and `attributeFlag` used to be required since the `GetPropertyValue()` and `GetMethodValue()` previously called `HasGettableProperty()` and `HasGettableMethod()` respectively. Those were removed as part of the refactoring to the underlying `MemberAccessor.GetValue()` (a43434dc), which now recreates just the aspects of the validation logic which are necessary for retrieving a value.

If it is a concern that the value be compatible with a particular `targetType` or `attributeFlag`, then callers can (and should) continue to call `HasGettableProperty()` and `HasGettableMethod()` to ensure compatibility. Otherwise, however, this is not the responsibility of `GetPropertyValue()` and `GetMethodValue()`, and doesn't affect their functionality.