Removed legacy `MemberInfoCollection`
The `MemberInfoCollection` offered a convenience shortcut to `MemberInfoCollection<T>` without needing to specify a `MemberInfo` type. This was never actually used—and with the focus being on e.g. `TypeAccessorCache`, it's unlikely it'll be used in the future.

For now, we're keeping the strongly typed `MemberInfoCollection<T>`, which can be used if there's a need for such a collection. (Though we'll likely revisit whether this is needed as well in the future.)

As part of this, I also fixed a couple of XML Doc references that were _intended_ to reference the generic version, but inadvertently references _this_ version.