Introduced `GetPrimaryConstructor()` convenience method
The legacy `MemberInfoCollection<T>` which `MemberDispatcher` relied upon included all members, including constructors, as part of its collection. With the migration to `MemberAccessor`, however, the `TypeAccessor` that replaces `MemberInfoCollection<T>` doesn't track constructors. This is because it doesn't have enough information to create delegates for constructors, nor do the semantics of `MemberAccessor` really work with constructors (e.g., `GetValue()`, `CanWrite`, &c.).

As part of the migration from `MemberDispatcher`,  the `TypeAccessor` inherited a `GetMembers<T>()` method that is exclusively used for getting constructors via the legacy `MemberInfoCollection<T>`. Since this is the only current need, a more elegant approach is to simply offer a purpose-built convenience method to retrieve the _primary_ constructor (currently defined as the first public, instance constructor).