(Re)Introduced `GetMembers<T>()` as a convenience method
In a previous commit, I got rid of the `GetMembers<T>()` extension (68f582d5) as it was attempting—poorly!—to satisfy legacy requirements based on an organization of `MemberInfo` instances grouped by `MemberInfoCollection<T>` collections. There remains a need to extract `MemberInfo` instances from a `TypeAccessor`, however, and optionally filter by type.

This is primarily used by the `BindingModelValidator`, which needs an enumeration of `PropertyInfo` objects. As part of this, I've retrofitted the two areas that relied on this to use the new overload. This satisfies the basic condition of the legacy extension method, but with much cleaner code, and without a dependency on the legacy `MemberInfoCollection<T>`.