Removed dependency on `MemberInfoCollection<T>` in tests
The `GetMembers_Property_ReturnsProperties` unit test, which now evaluates the new `GetMembers<T>()` overload (c85de0a9), was relying on `MemberInfoCollection<T>` not as a necessary subject of testing, but as a convenience for evaluating the results of a collection. This is mitigated by using XUnit's `Assert.Contains()` method instead.