Removed legacy `MemberInfoCollection<T>` class
With the last few commits, there's no longer any dependency on the legacy `MemberInfoCollection<T>` class. It does offer a potential _convenience_ as a keyed collection of `MemberInfo` objects, but as this isn't strictly necessary anymore—and, indeed, we're no longer actively using it—it doesn't make sense to keep it around, and especially as an internal class.

Should we, at a later date, rediscover a need for something like this, it's easy enough to resurrect—or simply recreate based on the precise needs at that time. Until then, this is the last vestigial element of the legacy `MemberDispatcher` and `TypeMemberInfoCollection` system that's now been replaced by `TypeAccessor` and `MemberAccessor`.

As part of this, removed a few legacy unit tests that were part of the `TypeAccessor` extensions test (a location they didn't really fit into anymore).