Consolidated logic into `GetValue()`, `SetValue()`
The `GetValue()` and `SetValue()` methods didn't get any business logic for e.g. validating parameters, and wouldn't throw an exception if called with an invalid member. This is inconsistent with other methods. In the meanwhile, the e.g. `GetPropertyValue()` and `GetMethodValue()` methods essentially shared the same logic. To mitigate both issues, the logic from the property- and method-based actions was moved into `GetValue()` and `SetValue()`.

A downside of this is that it necessitated removing the check for whether the `MemberAccessor` is a property or a method. This remains in e.g. `HasGettableProperty()`, if needed. I continue to question the need for the property- and method-based interface, however, and this helps deprioritize their priority by making them simple convenience passthroughs until a decision is made.