Centralized `IsSettable()` on `MemberAccessor`
This simply allows `IsSettable()` to be called directly, if need be, off of the `MemberAccessor` without relying on a reference to the `TypeAccessor`. This isn't currently needed, but it doesn't really add any code or complexity.

While I was at it, I improved the `IsSettable()` logic to be more consistent with the actual `GetValue()` and `SetValue()` logic. For instance, it now checks both source and target type convertibility, and optionally allows convertibility to be toggled. We're not exposing `allowConversion` on the hassable methods (e.g., `HasGetterProperty()`), so these are hard-coded to `true` in those calls.