FAILS: Attempt to capture `MapAsAttribute<>`
Because unbound types cannot be used as generic type arguments, we can't look for `MapAsAttribute<>`. To allow it to compile, we instead use `MapAsAttribute<object>`. This works—but, of course, fails to recognize any other `MapAsAttribute<>`. To fix this, we'd either need to make this so it doesn't rely on generics (which will be difficult with the delegate), or we'd need to dynamically assess the type and make the call via reflection (which is ugly).

This likely makes this feature useless for our needs, as it's a bit of work, a performance hit, and very little benefit.