Introduced global using directives for ASP.NET Core integration tests
Introduced global using directives for `OnTopic.AspNetCore.Mvc.IntegrationTests`

The three most commonly referenced namespaces in this project are, unsurprisingly:
- ` Microsoft.AspNetCore.Mvc.Testing`
- `OnTopic.AspNetCore.Mvc.IntegrationTests.Host`
- `Xunit`

These only impact three files—but that's also 100% of the files in this small project. These have been added as `global using` directives (in the `AssemblyInfo`) and removed from the individual files.