Removed `Xunit` namespace from global using directives
This is a bug in my previous commit (c9f6b179). In adding global using directives, the `Xunit` namespace introduced an ambiguity with the `[Collection()]` attribute, which is also used by `OnTopic`. As a result, I removed the `Xunit` namespace.

This isn't that big of a deal since, despite there being over twenty files with the `Xunit` namespace, there are far more files in the test project representing POCO objects for testing, and many of those use the `[Collection()]` attribute.

This is an amendment to #93. Unfortunately, this got missed as part of the merge (c9f6b179). Whoops.