Delete `Program` in preparation for ASP.NET Core 6.0 Hosting Model
In ASP.NET Core 6, the new Hosting Model allows for the `Program` and `Startup` files to be merged into one. While the final merged file will live at `Program.cs`, almost the entirety of the logic will be from the `Startup` class and, thus, we want to start from that foundation. In preparation for this, we're deleting the current `Program` class.