Move `Startup` to `Program` as part of ASP.NET Core 6.0 Hosting Model
In ASP.NET Core 6, the new Hosting Model allows for the `Program` and `Startup` files to be merged into one. Since almost the entirety of the logic will be from the `Startup` class, we want to start from that foundation. In preparation for this, we're moving `Startup` to `Program`, since that's where the app will ultimately be loaded from.