Elevated `Startup` to top-level statements
With the ASP.NET 6 Hosting Model, an explicit `Startup` or `Program` class aren't required; instead, all statements can be moved to C# 9.0 top-level statements. This reduces some of the ceremony and boilerplate around the application configuration, while also reducing the indentation.

As an initial step, the namespace, class, and methods are removed, but the indentation and variable names remain the same. This makes the diff more readable. In subsequent commits, I'll fix the indentation and standardize the variable names.