Established new `ErrorController`
We previously had an `ErrorController`, but it didn't provide any real intelligence, and didn't integrate well with either ASP.NET Core or OnTopic. As such, it was removed with OnTopic 4.

This new `ErrorController` fixes both issues. It is written to accommodate ASP.NET Core's e.g. `UseStatusCodePages()` by accepting an `errorCode` in the path. And it integrates with OnTopic by intelligently falling back to different topics. E.g., if a 403 error is submitted, it will first look for a topic at `Error:403`, then `Error:400`, and finally deliver the root `Error` topic.

This satisfies the core functionality of #91, though we'll still be committing updates to the documentation, unit tests, and integration tests—as well as a convenience extension method to simplify configuring routes.