Established new `MapTopicErrors()` extension method
The `MapTopicErrors()` extension method isn't _necessary_, but it _optionally_ allows the route `/Error/{statusCode}` to be used while still taking advantage of the `/Error/Http/{statusCode}` fallback logic (see e7fe64fa).

With this, the route can instead be configured using `MapDefaultControllerRoute()`, which will correctly route to `/Error/Http/{statusCode}`.

This adds a convenience method for configuring the `ErrorController`, as proposed in #91.