Introduced unit tests for `ErrorController`
Provides unit tests that evaluate the logic in the `ErrorController.HttpAsync()` method. Specifically ensures that the fallback logic works, by returning a) an exact match (if available), otherwise b) the parent category (if available), and finally c) the `Error` root (which is assumed to be available).

This resolves the unit testing portion of #91. We'll additionally be introducing integration tests in future commits.