Added minimal `Page` view
The `Page` view allows any arbitrary `Page` content type to be rendered by the integration tests. It will return, simply, the `Topic.Title`, thus making it easy to evaluate the result as part of the integration tests.

This is being done in anticipation of adding integration tests for the new `ErrorController` (e7fe64fa), though it may well prove useful for other integration tests in the future.