Extended integration test to validate routing to `ErrorController`
Extended the new `[Theory]` integration test (32306875) to validate that the various routes configured for the `ErrorController` (525f65b3) are, in fact, functioning as expected.

This includes the routes `/{rootTopic}/{errorCode}` route (from `MapTopicErrors()`), the `/{controller}/{action}` route (from `MapDefaultControllerRoute()`), and the `/{rootTopic}/{path}` route (from `MapTopicRoute()`).

This satisfies the first part of the integration tests for #91.