Updated `UseExceptionHandler()` to use the new `ErrorController`
With the introduction of the `ErrorController` (e7fe64fa, bf46877a), we now have a valid target for the `UseExceptionHandler()` call. Previously, this pointed to a non-existent page. This was fine for sample purposes, but now that we have an actual page to us, this is preferred.

This could alternatively point to any arbitrary `Topic` under `Error` (e.g., `Error:InternalServer`).