Establish a `CacheProfile` topic view model, if needed
It shouldn't be necessary to access the `CacheProfile` topic via views since the cache profile attributes are set via the new `[TopicResponseCache]`. However, should it be necessary (e.g., to update the meta headers), the `CacheProfileTopicViewModel` allows one to be returned to the client.

This isn't (currently) exposed via e.g. the `PageTopicViewModel` since it's not _expected_ to be needed, and we don't want to add potential overhead to _every_ page mapping.