Throw exception is `[TopicResponseCache]` not applied to `TopicController`
The `[TopicResponseCache]` will only work with the `TopicController`. Initially, this was validated without an exception since it was intended to be registered as a global filter. In the end, however, it made more sense to register it as an attribute filter on the `TopicController` itself, thus automatically registering it with all derived controllers. Given that, it should throw an explicit exception if a developer attempts to add it to any other type of controller. This maintains consistency with the `[ValidateTopic]` attribute, which performs the same validation.