Fixed specific type of `NotFoundResult` in unit test
When content is included in the `NotFound` result, the `Controller` class returns a `NotFoundObjectResult`. When no content is included, however, it returns a `NotFoundResult`. As such, when I upaded the `[ValidateTopic]` action filter to return `NotFound()` without a specific error message (ab71a692), I also needed to update the `ValidateTopicAttributeTest` to correctly expect the `NotFoundResult` instead of the `NotFoundObjectResult`. Whoops.