Prioritized `if`/`else if` statements in `GetValue()`
Previously, `GetValue()` was prioritized based on how common a given scenario was. Now, it's prioritized in part based on how expensive that check is. For instance, checking `[MapToParent]` is cheap, so there's no need to first check the comparatively more expensive `IsConvertible()` or `IsList()` first. This is a trivial optimization, but given how frequent this library is called, it's easy to make.