Ensure `IsConvertible()` isn't checked for `IsList()` or `getTopicReference()`
Previously, the `IsConvertible()` check was bypassed if `mapAssociationsOnly` was set. That makes sense, because there's no need to call `GetScalarValue()` in that case. Except that also necessitated that each of the subsequent conditions be checked, which collectively are more expensive than the check to `IsConvertible()`. As such, since scalar values are the most common type of property we're working with, it's cheaper to put this upfront and then check `mapAssociationsOnly` within it instead.