Only check for `null` within `GetScalarValue()`
If there's not a getter method or getter property, then `attributeValue` will be null. If there is a getter method or getter property, then `attributeValue` might be null or empty. If it's null, that's ambiguous between whether the method existed in not. But if it's empty, we know the member existed and returned a value. In that case, we should trust that value instead of executing fallback logic with `String.IsNullOrEmpty()`.

A further optimization would be to confirm that it exists and only fallback if it doesn't. As these are comparatively rare scenarios, though, it doesn't make sense to add additional overhead for this check, and especially when it's already handled by the `GetMethodValue()` and `GetPropertyValue()` methods.