Move expensive `SatisfiesAttributeFilters()` call to end of loop
The `SatisfiesAttributeFilters()` call should be reasonably fast, but it's expensive compared to the other validation checks in the `PopulateTargetCollectionAsync()` loop, and there's no need to call it if we can easily reject adding a topic because it violates the `ContentTypeFilter`, is a `List`, or `IsDisabled`.

In practice, the most significant impact we'd expect here would be if `[FilterByAttribute()]` is combined with `[FilterByContentType()]`. These used to be the same filter, but since the content type was moved out of attributes, a new filter attribute needed to be introduced. There's no sense in evaluating the attributes if the content type isn't valid.